pyproject.toml
[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
name = "portfolio_balancer"
version = "0.1.0"
description = "Simple portfolio balancing tool"
authors = ["Mike Conigliaro <mike@conigliaro.org>"]
homepage = "https://github.com/mconigliaro/portfolio_balancer"

[tool.poetry.dependencies]
python = ">=3.8,<3.10"

click = "^8.0"
pyyaml = "^ 5.4"
rich = "^10.12"

[tool.poetry.dev-dependencies]
pytest = "^6.2"
pytest-cov = "^3.0"
pytest-xdist = "^2.4"

[tool.poetry.scripts]
pb = 'portfolio_balancer.cli:main'

[tool.pyright]
venvPath = "."
venv = ".venv"

[tool.pytest.ini_options]
addopts = "--capture=tee-sys"
log_cli = 1
log_level = "DEBUG"
